Creating a new project structure based on maven-archetype-webapp. Moving objects to the proper location. Fixing issue in Luhn.generateCheckDigit. Updating JUnit for generateCheckDigit method after fixing. Fixing JUnit for isValidLuhn method.
