--- README.md
@@ -1,9 +1,0 @@
-# Instructions
-Use the [GitHub Flow](https://guides.github.com/introduction/flow/) or similar feature-branch (with merges) based development process to fix these [Issues](https://github.com/Chockstone/interview/issues?q=is%3Aissue+is%3Aopen+sort%3Acreated-asc) in your own fork of this repository.  You DO NOT have to create Pull Requests.
-
-Demonstrate that you can:
-* keep it simple: concise patches _for **each** Issue_that will be easy for the team to review
-* show your work: explain succintly in commits / comments what goes into large patches
-   * feel free to have work-in-progress branches along with final merge-ready (e.g. squashed) branches
-
-Upon completion, send an e-mail with the URL of your public fork.
