+++ card-validator/.gitignore
@@ -56,6 +56,8 @@ local.properties
 .cache-main
 .scala_dependencies
 .worksheet
+.project
+.classpath
 
 # JetBrains
 .idea/**/workspace.xml
