+++ card-validator/README.md
@@ -1,0 +1,8 @@
+# Card Validator Project
+This project was created to accepts a card number and determines if the card number is a valid number with respect to the Luhn algorithm.
+
+## Build procedure
+To build the project, use the following maven command:
+```cmd
+mvn clean install
+```
