+++ card-validator/src/main/java/com/hps/luhn/Luhn.java
@@ -7,13 +7,13 @@ public class Luhn {
 
 	/**
 	 * TODO
-	 * 
+	 *
 	 * Accepts a card number and determines if the card number is a valid number
 	 * with respect to the Luhn algorithm.
-	 * 
+	 *
 	 * @param cardNumber
 	 *            the card number
-	 * 
+	 *
 	 * @return true if the card number is valid according to the Luhn algorithm,
 	 *         false if not
 	 */
@@ -24,10 +24,10 @@ public class Luhn {
 	/**
 	 * Accepts a partial card number (excluding the last digit) and generates
 	 * the appropriate Luhn check digit for the number.
-	 * 
+	 *
 	 * @param cardNumber
 	 *            the card number (not including a check digit)
-	 * 
+	 *
 	 * @return the check digit
 	 */
 	public int generateCheckDigit(int cardNumber) {
@@ -35,14 +35,14 @@ public class Luhn {
 		int sum = 0;
 		while (cardNumber > 0) {
 			// starting from the right (rightmost is the unknown check digit)
-			long digit = cardNumber % 10; 
+			long digit = cardNumber % 10;
 
 			if (doubleDigit) { // double the value of every second digit
 				digit *= 2;
 
 				// if two digits, use the sum of the digits
-				if (digit > 10) {
-					digit = digit / 10 + digit % 10; 
+				if (digit >= 10) {
+					digit = digit / 10 + digit % 10;
 				}
 			}
 			doubleDigit = !doubleDigit;
@@ -57,16 +57,16 @@ public class Luhn {
 
 	/**
 	 * TODO
-	 * 
+	 *
 	 * Accepts two card numbers representing the starting and ending numbers of
 	 * a range of card numbers and counts the number of valid Luhn card numbers
 	 * that exist in the range, inclusive.
-	 * 
+	 *
 	 * @param startRange
 	 *            the starting card number of the range (may not be valid luhn)
 	 * @param endRange
 	 *            the ending card number of the range (may not be a valid luhn)
-	 * 
+	 *
 	 * @return the number of valid Luhn card numbers in the range, inclusive
 	 */
 	public int countRange(int startRange, int endRange) {
